new pikachu and charmeleon class, still has errors.

+++ PokeBattle/PokeBattle/Weakness.cs
@@ -9,7 +9,7 @@ namespace PokeBattle
         private string _energy;
         private float _multiply;
 
-        public Weakness(string _energy, int _multiply)
+        public Weakness(string _energy, float _multiply)
         {
             this._energy = _energy;
             this._multiply = _multiply;
